requirements.txt
annotated-types==0.7.0
anyio==3.7.1
Authlib==1.2.1
bidict==0.23.1
blake3==1.0.4
blinker==1.7.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.1.1
cppyy==3.0.0
cppyy-backend==1.14.11
cppyy-cling==6.28.0
CPyCppyy==1.12.13
cryptography==44.0.3
cycler==0.12.1
dnspython==2.7.0
fastapi==0.104.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.3.6
fonttools==4.43.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.9.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.25.1
idna==3.10
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.0
motor==3.3.2
mysqlclient==2.2.4
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.9.0
python-multipart==0.0.6
python-socketio==5.11.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scapy==2.5.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
utils==1.0.1
uvicorn==0.24.0
websockets==15.0.1
Werkzeug==3.0.2
wsproto==1.2.0
